Whisper Tracing v3
whisper tracing protocol enhancement. In this upgrade, whisper peers
share public key and compute unique interaction tokens per encounter.
For each interaction, each peer do the following:
- share public key g^a and receive peer public key g^b (Curve25519)
- compute share secret using diffie-hellman K=g^ab
- create two tokens by hashing K with both pub keys
  - tell token: H1 = Hash(K,g^a)
  - hear token: H2 = Hash(K,g^b)

The tell token is ought to be poublished if the current user tested
positive for infection. the hear token is used to be matched against the
public list of token. This can be done in a decentralized fashion by
downloading the list of public token and perform the matching locally,
or it can be done using a server to do the match-making.